# project/openapi.yaml
openapi: 3.0.3
info:
  title: 我的数字老师 - 后端接口
  version: 0.1.0
  description: |
    由当前 uni-app 前端项目整理出的"需要/预留"的后端接口规格。
    可直接导入 Swagger / Apifox / Postman。
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        ok:
          type: boolean
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "ok"
        data:
          nullable: true
...
    UploadResponseData:
      type: object
      properties:
        url:
          type: string
        mime:
          type: string
        size:
          type: integer
...
    AvatarProfile:
      type: object
      properties:
        age:
          type: integer
          minimum: 3
          maximum: 12
        interests:
          type: array
          items:
            type: string
        style:
          type: string
        freeText:
          type: string
    AvatarVoice:
      type: object
      properties:
        presetId:
          type: string
        note:
          type: string
    AvatarOutput:
      type: object
      properties:
        characterStyle:
          type: string
    AvatarGenerateRequest:
      type: object
      required: [portraitUrl, profile, voice]
      properties:
        portraitUrl:
          type: string
          description: 头像原图 URL（必须为 /files/...）
        profile:
          $ref: '#/components/schemas/AvatarProfile'
        voice:
          $ref: '#/components/schemas/AvatarVoice'
        output:
          $ref: '#/components/schemas/AvatarOutput'
    AvatarGenerateResponseData:
      type: object
      properties:
        avatarId:
          type: string
        imageUrl:
          type: string
        url:
          type: string
        scope:
          type: string
          enum: [account, user]
    AvatarActiveResponseData:
      type: object
      properties:
        avatarId:
          type: string
        imageUrl:
          type: string
        portraitUrl:
          type: string
        profile:
          $ref: '#/components/schemas/AvatarProfile'
        voice:
          $ref: '#/components/schemas/AvatarVoice'
        createdAt:
          type: integer
        scope:
          type: string
          enum: [account, user]
    AvatarVideoGenerateRequest:
      type: object
      properties:
        imageUrl:
          type: string
          description: 头像图片 URL（/files/..）
        introText:
          type: string
          description: 自我介绍文本（用于生成声音描述）
        prompt:
          type: string
          description: 视频提示词
        resolution:
          type: string
          enum: [720P, 1080P]
        duration:
          type: integer
          description: 秒数（建议 10~15）
        shotType:
          type: string
          description: 多镜头类型（multi 等）
    AvatarVideoGenerateResponseData:
      type: object
      properties:
        jobId:
          type: string
        audioUrl:
          type: string
    AvatarVideoStatusResponseData:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        videoUrl:
          type: string
        scope:
          type: string
          enum: [account, user]
paths:
...
  /upload/image:
    post:
      tags: [Upload]
      summary: 上传图片（聊天附件）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponseData'
...
  /upload/voice:
    post:
      tags: [Upload]
      summary: Upload voice recording
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /avatar/upload:
    post:
      tags: [Avatar]
      summary: 上传头像照片（数字人采集）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponseData'
  /avatar/generate:
    post:
      tags: [Avatar]
      summary: 生成数字人头像
      security:
        - bearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          description: 头像绑定范围（account=子账号；user=用户全局）
          schema:
            type: string
            enum: [account, user]
            default: account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarGenerateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AvatarGenerateResponseData'
  /avatar/active:
    get:
      tags: [Avatar]
      summary: 获取当前头像
      security:
        - bearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          description: 头像绑定范围（account=子账号；user=用户全局）
          schema:
            type: string
            enum: [account, user]
            default: account
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AvatarActiveResponseData'
  /avatar/video/generate:
    post:
      tags: [Avatar]
      summary: 生成头像视频（Wan2.6-I2V-Flash）
      security:
        - bearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          description: 头像绑定范围（account=子账号；user=用户全局）
          schema:
            type: string
            enum: [account, user]
            default: account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarVideoGenerateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AvatarVideoGenerateResponseData'
  /avatar/video/status:
    get:
      tags: [Avatar]
      summary: 查询头像视频生成状态
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: false
          description: 头像绑定范围（account=子账号；user=用户全局）
          schema:
            type: string
            enum: [account, user]
            default: account
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AvatarVideoStatusResponseData'